pyTab Project: main program and tab class
A new `tab.print()` class method was implemented to handled a limited
output for the tab that is desirable when the other class methods want
to show their effect on the tab data. Calling `print(tab_instance)` will
print the entire tab data with appropriate formatting. A way to handle
this in the terminal should be implemented in the main program.

A few minor edits were also made to the main program. Crucially, a bug
was caught in the argument parsing that meant the chord writing case was
incorrectly called if no input was given. This is related to the default
argument that is given for args.chord in the argument parser, leading to
a truth value in an if statement when there shouldn't be. This behaviour
was changed so the default of args.chord is now None, but more thought
needs to be given to this.